Fix precise probability quantization.
The key to the heap should be the incremental error for adjusting that
probability. cc should track the current quantized error which means
adding the incremental error to the previous error. The newly calculated
error then should then remove the previously accounted for error to get
the next incremental error.

for:
ANS test=init_power(0.11,8);
test.quantize_prec(6);
The quantization rate loss decreases from 0.201408 to 0.199105.
The rate loss for spread_tuned_s() decreases from 0.143563 to 0.139374.